Refactor Registry
