+++ lib/chronicle/etl/cli/plugins.rb
@@ -16,7 +16,7 @@ module Chronicle
           cli_fail(message: "Please specify a plugin to install") unless plugins.any?
 
           installed, not_installed = plugins.partition do |plugin|
-            Chronicle::ETL::Registry::PluginRegistry.installed?(plugin)
+            Chronicle::ETL::Registry::Plugins.installed?(plugin)
           end
 
           puts "Already installed: #{installed.join(", ")}" if installed.any?
@@ -27,7 +27,7 @@ module Chronicle
 
           not_installed.each do |plugin|
             spinner.update(title: "Installing #{plugin}")
-            Chronicle::ETL::Registry::PluginRegistry.install(plugin)
+            Chronicle::ETL::Registry::Plugins.install(plugin)
 
           rescue Chronicle::ETL::PluginError => e
             spinner.error("Error".red)
@@ -41,7 +41,7 @@ module Chronicle
         def uninstall(name)
           spinner = TTY::Spinner.new("[:spinner] Uninstalling plugin #{name}...", format: :dots_2)
           spinner.auto_spin
-          Chronicle::ETL::Registry::PluginRegistry.uninstall(name)
+          Chronicle::ETL::Registry::Plugins.uninstall(name)
           spinner.success("(#{'successful'.green})")
         rescue Chronicle::ETL::PluginError => e
           spinner.error("Error".red)
@@ -51,7 +51,7 @@ module Chronicle
         desc "list", "Lists available plugins"
         # Display all available plugins that chronicle-etl has access to
         def list
-          values = Chronicle::ETL::Registry::PluginRegistry.all
+          values = Chronicle::ETL::Registry::Plugins.all
             .map do |plugin|
             [
               plugin.name,
